requirements.txt
playwright==1.57.0
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.4
